==> app/models/step_2.ipynb <==
# Курсовой проект по курсу "Машинное обучение в бизнесе".

## Соковнин Игорь Леонидович

### 2 шаг. Проверка работоспособности и качества пайплайна

Здесь мы еще не запускаем никакое API, а загружаем модель (pipeline) напрямую и проверяем на отложенной (тестовой) выборке

import pandas as pd
from sklearn.metrics import roc_auc_score,roc_curve # ,scorer
import dill
dill._dill._reverse_typemap['ClassType'] = type

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

X_test.head(3)

columns = ['Purpose_int', 'Term_int', 'Tax_Liens_int', 'Number_of_Credit_Problems_int']

Purpose_int, Term_int, Tax_Liens_int, Number_of_Credit_Problems_int = "", "", "", ""

Purpose_int = "14"
Term_int = "1"
Tax_Liens_int = "0"
Number_of_Credit_Problems_int = "0"

with open('logreg_pipeline_v1.dill', 'rb') as in_strm:
    pipeline = dill.load(in_strm)

# with open('catboost_pipeline.dill', 'rb') as in_strm:
#     pipeline = dill.load(in_strm)

predictions = pipeline.predict_proba(pd.DataFrame({"Purpose_int": [Purpose_int],
                                  "Term_int": [Term_int],
                                  "Tax_Liens_int": [Tax_Liens_int],
                                  "Number_of_Credit_Problems_int": [Number_of_Credit_Problems_int]}))
predictions

columns = ['Annual Income', 'Number of Open Accounts', 'Years of Credit History',
       'Maximum Open Credit', 'Current Loan Amount', 'Current Credit Balance',
       'Monthly Debt', 'CreditScore_small', 'CreditScore_large',
       'Home_Ownership_int', 'Years_in_current_job_int', 'Purpose_int',
       'Term_int', 'Tax_Liens_int', 'Number_of_Credit_Problems_int']

predictions = pipeline.predict_proba(X_test[columns])
pd.DataFrame({'preds': predictions[:, 1]}).to_csv("test_predictions.csv", index=None)
predictions

roc_auc_score(y_score=predictions[:, 1][:], y_true=y_test.iloc[:])
