==> app/models/step_3_15.ipynb <==
### Проверка того, что наше api работает

Перед началом необходимо собрать docker-образ и запустить контейнер

1. Следовать инструкциям https://github.com/fimochka-sudo/GB_docker_flask_example/blob/main/README.md
2. Далее уже запускать код ниже

import pandas as pd
from sklearn.metrics import roc_auc_score,roc_curve,scorer
from urllib import request, parse

X_test = pd.read_csv("X_test.csv")
y_test = pd.read_csv("y_test.csv")

# X_test.columns
columns = ['Annual Income', 'Number of Open Accounts', 'Years of Credit History',
       'Maximum Open Credit', 'Current Loan Amount', 'Current Credit Balance',
       'Monthly Debt',
       'CreditScore_small', 'CreditScore_large',
       'Home_Ownership_int', 'Years_in_current_job_int', 'Purpose_int',
       'Term_int', 'Tax_Liens_int', 'Number_of_Credit_Problems_int']

# X_test[columns].head(3)

import urllib.request
import json      

def get_prediction(Annual_Income,
                   
                   Number_of_Open_Accounts,
                   Years_of_Credit_History,
                   Maximum_Open_Credit,
                   Current_Loan_Amount,
                   Current_Credit_Balance,
                   Monthly_Debt,
                   CreditScore_small,
                   CreditScore_large,
                   
                   Home_Ownership_int,
                   Years_in_current_job_int,
                   Purpose_int,
                   Term_int,
                   Tax_Liens_int,
                   Number_of_Credit_Problems_int):
    
                    # Annual Income : 0.616627
                    # Number_of_Open_Accounts : 0.380774
                    # Years_of_Credit_History : -0.345022
                    # Maximum_Open_Credit : 0.41448
                    # Current_Loan_Amount : -0.506407
                    # Current_Credit_Balance : 0.775597
                    # Monthly_Debt : 0.342168
                    # CreditScore_small : 0.381853
                    # CreditScore_large : -0.237124
                    # Home_Ownership_int : 3
                    # Years_in_current_job_int : 8
                    # Purpose_int : 14
                    # Term_int : 1
                    # Tax_Liens_int : 0
                    # Number_of_Credit_Problems_int : 0


    body = {'Annual Income': str(Annual_Income),

            'Number of Open Accounts': str(float(Number_of_Open_Accounts)),
            'Years of Credit History': str(float(Years_of_Credit_History)), 
            'Maximum Open Credit': float(float(Maximum_Open_Credit)), 
            'Current Loan Amount': float(float(Current_Loan_Amount)), 
            'Current Credit Balance': str(float(Current_Credit_Balance)),

            'Monthly Debt': str(float(Monthly_Debt)),
            'CreditScore_small': str(float(CreditScore_small)),
            'CreditScore_large': str(float(CreditScore_large)),
            
            'Home_Ownership_int': str(int(Home_Ownership_int)),
            'Years_in_current_job_int': str(int(Years_in_current_job_int)), 
            'Purpose_int': str(int(Purpose_int)), 
            'Term_int': str(int(Term_int)), 
            'Tax_Liens_int': str(int(Tax_Liens_int)),
            'Number_of_Credit_Problems_int': str(int(Number_of_Credit_Problems_int))
            } 
    
#     for x in body:
#         print (x,':',body[x])
    
    jsondata = json.dumps(body)
    jsondataasbytes = jsondata.encode('utf-8')   # needs to be bytes
    # print (jsondataasbytes)
    # myurl = "http://0.0.0.0:8180/predict"
    myurl = "http://0.0.0.0:8182/predict"
    req = urllib.request.Request(myurl)
    req.add_header('Content-Type', 'application/json; charset=utf-8')
    req.add_header('Content-Length', len(jsondataasbytes))

    response = urllib.request.urlopen(req, jsondataasbytes)

    return json.loads(response.read())['predictions']

#  get_prediction(0.616627, 3, 8, 14, 1, 0, 0)

X_test[columns].head(1)

get_prediction(0.616627,
               0.380774,
               -0.345022,
               0.414480,
               -0.506407,
               0.775597,
               0.342168,
               0.381853,
               -0.237124,
               3,
               8,
               14,
               1,
               0,
               0)

%%time
predictions = X_test.iloc[:100].apply(lambda x: get_prediction(x[0], 
                                                             x[1],
                                                             x[2],
                                                             x[3],
                                                             x[4],
                                                             x[5],
                                                             x[6],
                                                             x[7],
                                                             x[8],
                                                             x[9],
                                                             x[10],
                                                             x[11],
                                                             x[12],
                                                             x[13],
                                                             x[14]), 1)

roc_auc_score(y_score=predictions.values, y_true=y_test.iloc[:100])

Давайте подумаем, где наше api может сломаться

# get_prediction("", "", None)
